deprecate ThemeVisualDensity in favor of VisualDensity
