+++ sdk/python/packages/flet-core/src/flet_core/types.py
@@ -101,15 +101,15 @@ OptionalString = Optional[str]
 class MaterialStateDeprecated(EnumMeta):
     def __getattribute__(self, item):
         if item in [
-            "hovered",
-            "focused",
-            "pressed",
-            "dragged",
-            "selected",
-            "scrolledUnder",
-            "disabled",
-            "error",
-            "",
+            "HOVERED",
+            "FOCUSED",
+            "PRESSED",
+            "DRAGGED",
+            "SELECTED",
+            "SCROLLED_UNDER",
+            "DISABLED",
+            "ERROR",
+            "DEFAULT",
         ]:
             warn(
                 "MaterialState enum is deprecated and will be removed in version 0.26.0. "
@@ -358,6 +358,38 @@ class StrokeJoin(Enum):
     MITER = "miter"
     ROUND = "round"
     BEVEL = "bevel"
+
+
+class ThemeVisualDensityDeprecated(EnumMeta):
+    def __getattribute__(self, item):
+        if item in [
+            "STANDARD",
+            "COMPACT",
+            "COMFORTABLE",
+            "ADAPTIVE_PLATFORM_DENSITY",
+        ]:
+            warn(
+                "ThemeVisualDensity enum is deprecated and will be removed in version 0.27.0. "
+                "Use VisualDensity enum instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+        return EnumMeta.__getattribute__(self, item)
+
+
+class ThemeVisualDensity(Enum, metaclass=ThemeVisualDensityDeprecated):
+    STANDARD = "standard"
+    COMPACT = "compact"
+    COMFORTABLE = "comfortable"
+    ADAPTIVEPLATFORMDENSITY = "adaptivePlatformDensity"
+    ADAPTIVE_PLATFORM_DENSITY = "adaptivePlatformDensity"
+
+
+class VisualDensity(Enum):
+    STANDARD = "standard"
+    COMPACT = "compact"
+    COMFORTABLE = "comfortable"
+    ADAPTIVE_PLATFORM_DENSITY = "adaptivePlatformDensity"
 
 
 # Events
